Feat/saga pattern implementation

+++ services/payment_service/app/models/customer_wallet.py
@@ -9,7 +9,7 @@ class UserWallet(Base):
     
     id: Mapped[uuid.UUID] = mapped_column(primary_key=True, default=uuid.uuid4)
     customer_id: Mapped[uuid.UUID] = mapped_column(nullable=False, unique=True, index=True)
-    initial_amount: Mapped[int] = mapped_column(Integer, nullable=False, default=100000)
-    current_balance: Mapped[int] = mapped_column(Integer, nullable=False, default=100000)
+    initial_amount: Mapped[float] = mapped_column(Integer, nullable=False, default=100000)
+    current_balance: Mapped[float] = mapped_column(Integer, nullable=False, default=100000)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=func.now())
     updated_at: Mapped[datetime] = mapped_column(DateTime, default=func.now(), onupdate=func.now())
